new TripalViewsHooks class
