+++ tripal/tripal.services.yml
@@ -11,10 +11,6 @@ services:
   tripal.fileretriever:
     class: 'Drupal\tripal\Services\TripalFileRetriever'
     arguments: ['@http_client', '@tripal.logger']
-  tripal.hooks:
-    class: 'Drupal\tripal\Hook\TripalHooks'
-  tripal.theme:
-    class: 'Drupal\tripal\Hook\TripalTheme'
   tripal.logger:
     class: 'Drupal\tripal\Services\TripalLogger'
   tripal.rebuild_service:
@@ -76,6 +72,12 @@ services:
     arguments:
      - '@entity_type.manager'
      - '@database'
+  tripal.hooks:
+    class: 'Drupal\tripal\Hook\TripalHooks'
   tripal.entity_hooks:
     class: 'Drupal\tripal\Hook\TripalEntityHooks'
     autowire: TRUE
+  tripal.views_hooks:
+    class: 'Drupal\tripal\Hook\TripalViewsHooks'
+  tripal.theme:
+    class: 'Drupal\tripal\Hook\TripalTheme'
